changed category color type from string to int; added some mappers

+++ app/src/main/java/com/jk/financehelper/utils/DataUtils.kt
@@ -26,7 +26,7 @@ class DataUtils {
                         CategoryUI(
                             id = "x1",
                             "Sport",
-                            color = 0xFF0094C6
+                            color = (0xFF0094C6).toInt()
                         )
                     ),
                     scheduleList = listOf(),
@@ -49,7 +49,7 @@ class DataUtils {
                         CategoryUI(
                             id = "x1",
                             "Sport",
-                            color = 0xFF0094C6
+                            color =  (0xFF0094C6).toInt()
                         )
                     ),
                     scheduleList = listOf(),
